docs: 주석 수정
