+++ iOS/Projects/Core/Network/Tests/InterceptorTest.swift
@@ -15,12 +15,14 @@ final class InterceptorTest: XCTestCase {
   let interceptor = TestInterceptor()
 
   struct TestInterceptor: TNRequestInterceptor {
+    /// adapt함수를통해서 데이터를 받기 전, url를 "adapt"로 바꿉니다.
     func adapt(_ request: URLRequest, session _: URLSessionProtocol) -> URLRequest {
-      var urlrequest = request
-      urlrequest.url = URL(string: "adapt")!
-      return urlrequest
+      var request = request
+      request.url = URL(string: "adapt")!
+      return request
     }
 
+    /// 스테이터스코드가 401일때, statusCode가 200인 response로 바꾸고 data를 "retry".data(using: .utf8)로 바꿉니다.
     func retry(_ request: URLRequest, session _: URLSessionProtocol, data: Data, response: URLResponse, delegate _: URLSessionDelegate?) async throws -> (Data, URLResponse) {
       guard let statusCode = (response as? HTTPURLResponse)?.statusCode else {
         throw TNError.unknownError
@@ -44,9 +46,10 @@ final class InterceptorTest: XCTestCase {
     provider = nil
   }
 
-  func test_request가_adapt를했을때() async throws {
+  func test_request에_adapt를했을때() async throws {
     // arrange
     let tokkenDidBeExpierResponse = HTTPURLResponse(url: URL(string: "www.naver.com")!, statusCode: 200, httpVersion: nil, headerFields: nil)!
+    // url이 adapt가 아닌 이상 data는 "나는목이다"가 나오게 됩니다.
     mockSession = MockURLSession(
       mockData: "나는목이다".data(using: .utf8)!,
       mockResponse: tokkenDidBeExpierResponse,
@@ -62,7 +65,7 @@ final class InterceptorTest: XCTestCase {
     XCTAssertEqual(stringData, "아아답트")
   }
 
-  func test_request가_retry를했을때() async throws {
+  func test_request에_sessionResponse를401로설정한다면_retry에서_data를_retry로넘겨줍니다() async throws {
     // arrange
     let tokkenDidBeExpierResponse = HTTPURLResponse(url: URL(string: "www.naver.com")!, statusCode: 401, httpVersion: nil, headerFields: nil)!
     mockSession = MockURLSession(mockData: "나는목이다".data(using: .utf8)!, mockResponse: tokkenDidBeExpierResponse)
